Removed categories context and moved it to React-redux
