+++ src/App.js
@@ -1,11 +1,15 @@
+// React & libraries imports
 import { Routes, Route } from "react-router-dom";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
+//Utils imports
 import {
   onAuthStateChangedListener,
   createUserDocumentFromAuth,
 } from "./utils/firebase/firebase.utils";
+//Reducer Actions
 import { setCurrentUser } from "./store/user/user.action";
+//Component imports
 import Home from "./routes/home/home.component";
 import Navigation from "./routes/navigation/navigation.component";
 import Shop from "./routes/shop/shop.component";
