+++ src/routes/categories-preview/categories-preview.component.jsx
@@ -1,9 +1,10 @@
-import { useContext, Fragment } from "react";
-import { CategoriesContext } from "../../context/categories.context";
+import { Fragment } from "react";
+import { useSelector } from "react-redux";
+import { getCategories } from "../../store/categories/category.selector";
 import CategoryPreview from "../../components/category-preview/category-preview.component";
 
 const CategoriesPreview = () => {
-  const { categoriesMap } = useContext(CategoriesContext);
+  const categoriesMap = useSelector(getCategories);
 
   return (
     <Fragment>
