+++ src/routes/category/category.component.jsx
@@ -1,5 +1,6 @@
-import { useContext, useState, useEffect, Fragment } from "react";
-import { CategoriesContext } from "../../context/categories.context";
+import { useState, useEffect, Fragment } from "react";
+import { useSelector } from "react-redux";
+import { getCategories } from "../../store/categories/category.selector";
 import { useParams, Link } from "react-router-dom";
 import ProductCard from "../../components/product-card/product-card.component";
 import {
@@ -11,7 +12,7 @@ import {
 
 const Category = () => {
   const { category } = useParams();
-  const { categoriesMap } = useContext(CategoriesContext);
+  const categoriesMap = useSelector(getCategories);
   const [products, setProducts] = useState(categoriesMap[category]);
 
   useEffect(() => {
