--- src/routes/shop/DELETE.shop.styles.jsx
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-const ProductsContainer = styled;
-
-// .products-container{
-//   display: grid;
-//   grid-template-columns: repeat(4, 1fr);
-//   column-gap: 10px;
-//   row-gap: 50px;
-// }
