+++ src/store/categories/category.selector.js
@@ -1,0 +1,1 @@
+export const getCategories = (state) => state.categories.categoriesMap;
